requirements.txt
PyPDF2
python-docx
reportlab
sqlalchemy
celery
redis
openai
